make rvl/nand relocatable
