+++ pack/dol_objects.txt
@@ -245,21 +245,25 @@ out/rvlMemExpHeap.o
 out/rvlMemFrmHeap.o
 out/rvlMemUnitHeap.o
 out/dol/bss_80346d18_803481b0.o
-out/dol/sbss_8038683c_803868a0.o
+out/dol/sbss_8038683c_80386840.o
 out/dol/sdata2_803884a4_80388860.o
 out/rvlMemAllocator.o
 out/rvlMemList.o
 out/rvlMtx.o
 out/rvlMtx2.o
 out/rvlVec.o
-out/dol/sdata_80385a10_80385a98.o
 out/dol/sdata2_803888b4_803888b8.o
 out/rvlQuat.o
-out/nand.o
+out/rvlNand1.o
+out/rvlNand2.o
+out/dol/rodata_80252c84_80252dd0.o
+out/dol/sbss_80386844_80386848.o
+out/rvlNand3.o
 out/dol/text_8019f1a8_801a05b8.o
+out/dol/sdata_80385a64_80385a98.o
+out/dol/sbss_80386858_803868a0.o
 out/osAlarm.o
 out/osAlloc.o
-out/dol/rodata_80252c84_80252dd0.o
 out/dol/data_8028f068_8028f320.o
 out/dol/sdata_80385a9c_80385aa0.o
 out/osArena.o
